[#111]; fix: both modified problem

+++ src/main/kotlin/com/yourssu/soongpt/domain/timetable/implement/CourseCandidateFactory.kt
@@ -15,7 +15,8 @@ class CourseCandidateFactory {
         val courseTimes = CourseTimes.from(course.scheduleRoom)
         val compressTimes = compressTimes(courseTimes)
         val timeSlot = generateTimeSlot(compressTimes)
-        val point = course.point.toInt()
+        val point = course.point.toDouble().toInt()
+
         return CourseCandidate.from(course.code, timeSlot, point)
     }
